11.25.16 - Change Log
Scene
- Adusted minutes from Int to Double

GameTimeControl
- Formated newTime to a double with only 2 decimal points. Used DecimalFormat.

Map.java
- fixed some properties. Moved various movePlayer overloaded functions together. 
- fixed bugs. When calling map creation. 


MapControl
- Fixed bugs and uppdated so that its really moving our player arounds
- For each scene adusted 5 mintes to double 0.15

MapView
- Fixed issue so that the current location is now shown with a > <
- Added and commented out several lines of debugging code. 

StartProgramView
- Fixed bug where the Player was not actually being set into the WhereIsMyBone


Location
- Moved constructors to the top.
- Attempting to set location default isvisited to false. Something is still buggy here. When map is displayed all locations visited return true.

GameControl
- Fixed bug with setting player's initial location under CreateNewGame

+++ WhereIsMyBone/src/byui/cit260/whereismybone/model/Scene.java
@@ -23,7 +23,7 @@ public class Scene implements Serializable {
     private String name;        //name of current scene
     private Item item;          //perhaps null, if there is an item, need dialog
     public String mapSymbol;    //already visted the spot
-    public int minute;          //timed per location
+    public Double minute;          //timed per location
     
     public Scene(){ 
         this.description = "";
@@ -131,10 +131,10 @@ public class Scene implements Serializable {
         this.description = description;
     }
     
-    public int getMinute() {
+    public Double getMinute() {
         return minute;
     }
-    public void setMinute(int minute) {
+    public void setMinute(Double minute) {
         this.minute = minute;
     }
